Bump to v2 and add board layout
This design is intended for using a Raspberry Pi W and adds a header to
connect the Raspberry pi directly to the board. This also adds a PCB
speaker along with a transistor to amplify it, and a switching power
supply (prefab module). The transistor for the magnets is replaced by a
MOSFET (which has better specs and mostly a cuter package than the
TIP12x that was on there).